bird-migration-gcp/requirements.txt
google-cloud-bigquery
google-api-python-client
gunicorn==19.9.0
pandas